split loader and make investments work again
migration should now be complete in theory, optimizations left though
