+++ rpc/users.go
@@ -153,6 +153,7 @@ func validateUser() {
 		var prepRecipient opensolar.Recipient
 		var prepEntity opensolar.Entity
 
+		log.Println("PREPUSER: ", prepUser)
 		var x ValidateParams
 
 		prepInvestor, err = opensolar.RetrieveInvestor(prepUser.Index)
@@ -178,7 +179,6 @@ func validateUser() {
 			erpc.MarshalSend(w, x)
 			return
 		}
-
 	})
 }
