+++ testdata.go
@@ -1,49 +1,5 @@
 package main
 
-import (
-	"github.com/pkg/errors"
-	"log"
-
-	// edb "github.com/Varunram/essentials/database"
-	utils "github.com/Varunram/essentials/utils"
-	opensolar "github.com/YaleOpenLab/opensolar/core"
-	consts "github.com/YaleOpenLab/openx/consts"
-	database "github.com/YaleOpenLab/openx/database"
-	// opzones "github.com/YaleOpenLab/openx/platforms/ozones"
-	sandbox "github.com/YaleOpenLab/openx/sandbox"
-)
-
-func testSolarProject(index int, panelsize string, totalValue float64, location string, moneyRaised float64,
-	metadata string, invAssetCode string, debtAssetCode string, pbAssetCode string, years int, recpIndex int,
-	contractor opensolar.Entity, originator opensolar.Entity, stage int, pbperiod int, auctionType string) (opensolar.Project, error) {
-
-	var project opensolar.Project
-	project.Index = index
-	project.PanelSize = panelsize
-	project.TotalValue = totalValue
-	project.State = location
-	project.MoneyRaised = moneyRaised
-	project.Metadata = metadata
-	project.InvestorAssetCode = invAssetCode
-	project.DebtAssetCode = debtAssetCode
-	project.PaybackAssetCode = pbAssetCode
-	project.DateInitiated = utils.Timestamp()
-	project.EstimatedAcquisition = years
-	project.RecipientIndex = recpIndex
-	project.ContractorIndex = contractor.U.Index
-	project.OriginatorIndex = originator.U.Index
-	project.Stage = stage
-	project.PaybackPeriod = pbperiod
-	project.AuctionType = auctionType
-	project.InvestmentType = "munibond"
-
-	err := project.Save()
-	if err != nil {
-		return project, errors.New("Error inserting project into db")
-	}
-	return project, nil
-}
-
 /*
 // newLivingUnitCoop creates a new living unit coop
 func newLivingUnitCoop(mdate string, mrights string, stype string, intrate float64, rating string,
@@ -105,157 +61,41 @@ func newConstructionBond(mdate string, stype string, intrate float64, rating str
 	err = cBond.Save()
 	return cBond, err
 }
+
+/*
+	_, err = newConstructionBond("Dec 21 2021", "Security Type 1", 5.4, "AAA", "Moody's Investments", "Wells Fargo",
+		200000, "Opportunity Zone Construction", 200, "5% tax for 10 years", 1, "India Basin Project", "San Francisco", "India Basin is an upcoming creative project based in San Francisco that seeks to host innovators from all around the world")
+	if err != nil {
+		return err
+	}
+
+	_, err = newConstructionBond("Apr 2 2025", "Security Type 2", 3.6, "AA", "Ant Financial", "People's Bank of China",
+		50000, "Opportunity Zone Construction", 400, "No tax for 20 years", 1, "Shenzhen SEZ Development", "Shenzhen", "Shenzhen SEZ Development seeks to develop a SEZ in Shenzhen to foster creation of manufacturing jobs.")
+	if err != nil {
+		return err
+	}
+
+	_, err = newConstructionBond("Jul 9 2029", "Security Type 3", 4.2, "BAA", "Softbank Corp.", "Bank of Japan",
+		150000, "Opportunity Zone Construction", 100, "3% Tax for 5 Years", 1, "Osaka Development Project", "Osaka", "This Project seeks to develop cutting edge technologies in Osaka")
+	if err != nil {
+		return err
+	}
+
+	_, err = newLivingUnitCoop("Dec 21 2021", "Member Rights Link", "Security Type 1", 5.4, "AAA", "Moody's Investments", "Wells Fargo",
+		200000, "Coop Model", 4000, "India Basin Project", "San Francisco", "India Basin is an upcoming creative project based in San Francisco that seeks to host innovators from all around the world")
+	if err != nil {
+		return err
+	}
+
+	_, err = newLivingUnitCoop("Apr 2 2025", "Member Rights Link", "Security Type 2", 3.6, "AA", "Ant Financial", "People's Bank of China",
+		50000, "Coop Model", 1000, "Shenzhen SEZ Development", "Shenzhen", "Shenzhen SEZ Development seeks to develop a SEZ in Shenzhen to foster creation of manufacturing jobs.")
+	if err != nil {
+		return err
+	}
+
+	_, err = newLivingUnitCoop("Jul 9 2029", "Member Rights Link", "Security Type 3", 4.2, "BAA", "Softbank Corp.", "Bank of Japan",
+		150000, "Coop Model", 2000, "Osaka Development Project", "Osaka", "ODP seeks to develop cutting edge technologies in Osaka and invites investors all around the world to be a part of this new age")
+	if err != nil {
+		return err
+	}
 */
-
-// InsertDummyData inserts sample data
-func InsertDummyData(simulate bool) error {
-	var err error
-	// populate database with dummy data
-	var recp opensolar.Recipient
-	// simulate only if the bool is set to true. Simulates investment for three projects based on the presentation
-	// at the Spring Members' Week Demo 2019
-	if simulate {
-		if !consts.Mainnet {
-			log.Println("creating sandbox")
-			return sandbox.CreateSandbox()
-		} else {
-			return errors.New("you're on mainnet, not on a sandbox")
-		}
-	}
-	allRecs, err := opensolar.RetrieveAllRecipients()
-	if err != nil {
-		return err
-	}
-	if len(allRecs) == 0 {
-		// there is no recipient right now, so create a dummy recipient
-		var err error
-		recp, err = opensolar.NewRecipient("martin", "p", "x", "Martin")
-		if err != nil {
-			return err
-		}
-		recp.U.Notification = true
-		err = recp.U.AddEmail("varunramganesh@gmail.com")
-		if err != nil {
-			return err
-		}
-	}
-
-	var inv opensolar.Investor
-	allInvs, err := opensolar.RetrieveAllInvestors()
-	if err != nil {
-		return err
-	}
-	if len(allInvs) == 0 {
-		var err error
-		inv, err = opensolar.NewInvestor("john", "p", "x", "John")
-		if err != nil {
-			return err
-		}
-		err = inv.ChangeVotingBalance(100000)
-		// this function saves as well, so there's no need to save again
-		if err != nil {
-			return err
-		}
-		err = database.AddInspector(inv.U.Index)
-		if err != nil {
-			return err
-		}
-		x, err := database.RetrieveUser(inv.U.Index)
-		if err != nil {
-			return err
-		}
-		inv.U = &(x)
-		err = inv.Save()
-		if err != nil {
-			return err
-		}
-		err = x.Authorize(inv.U.Index)
-		if err != nil {
-			return err
-		}
-		inv.U.Notification = true
-		err = inv.U.AddEmail("varunramganesh@gmail.com")
-		if err != nil {
-			return err
-		}
-	}
-
-	originator, err := opensolar.NewOriginator("samuel", "p", "x", "Samuel L. Jackson", "ABC Street, London", "I am an originator")
-	if err != nil {
-		return err
-	}
-
-	contractor, err := opensolar.NewContractor("sam", "p", "x", "Samuel Jackson", "14 ABC Street London", "This is a competing contractor")
-	if err != nil {
-		return err
-	}
-
-	/*
-		_, err = newConstructionBond("Dec 21 2021", "Security Type 1", 5.4, "AAA", "Moody's Investments", "Wells Fargo",
-			200000, "Opportunity Zone Construction", 200, "5% tax for 10 years", 1, "India Basin Project", "San Francisco", "India Basin is an upcoming creative project based in San Francisco that seeks to host innovators from all around the world")
-		if err != nil {
-			return err
-		}
-
-		_, err = newConstructionBond("Apr 2 2025", "Security Type 2", 3.6, "AA", "Ant Financial", "People's Bank of China",
-			50000, "Opportunity Zone Construction", 400, "No tax for 20 years", 1, "Shenzhen SEZ Development", "Shenzhen", "Shenzhen SEZ Development seeks to develop a SEZ in Shenzhen to foster creation of manufacturing jobs.")
-		if err != nil {
-			return err
-		}
-
-		_, err = newConstructionBond("Jul 9 2029", "Security Type 3", 4.2, "BAA", "Softbank Corp.", "Bank of Japan",
-			150000, "Opportunity Zone Construction", 100, "3% Tax for 5 Years", 1, "Osaka Development Project", "Osaka", "This Project seeks to develop cutting edge technologies in Osaka")
-		if err != nil {
-			return err
-		}
-
-		_, err = newLivingUnitCoop("Dec 21 2021", "Member Rights Link", "Security Type 1", 5.4, "AAA", "Moody's Investments", "Wells Fargo",
-			200000, "Coop Model", 4000, "India Basin Project", "San Francisco", "India Basin is an upcoming creative project based in San Francisco that seeks to host innovators from all around the world")
-		if err != nil {
-			return err
-		}
-
-		_, err = newLivingUnitCoop("Apr 2 2025", "Member Rights Link", "Security Type 2", 3.6, "AA", "Ant Financial", "People's Bank of China",
-			50000, "Coop Model", 1000, "Shenzhen SEZ Development", "Shenzhen", "Shenzhen SEZ Development seeks to develop a SEZ in Shenzhen to foster creation of manufacturing jobs.")
-		if err != nil {
-			return err
-		}
-
-		_, err = newLivingUnitCoop("Jul 9 2029", "Member Rights Link", "Security Type 3", 4.2, "BAA", "Softbank Corp.", "Bank of Japan",
-			150000, "Coop Model", 2000, "Osaka Development Project", "Osaka", "ODP seeks to develop cutting edge technologies in Osaka and invites investors all around the world to be a part of this new age")
-		if err != nil {
-			return err
-		}
-	*/
-
-	_, err = testSolarProject(1, "100 1000 sq.ft homes each with their own private spaces for luxury", 14000, "India Basin, San Francisco",
-		0, "India Basin is an upcoming creative project based in San Francisco that seeks to invite innovators from all around to participate", "", "", "",
-		3, recp.U.Index, contractor, originator, 4, 2, "blind")
-
-	if err != nil {
-		return err
-	}
-
-	_, err = testSolarProject(2, "180 1200 sq.ft homes in a high rise building 0.1mi from Kendall Square", 30000, "Kendall Square, Boston",
-		0, "Kendall Square is set in the heart of Cambridge and is a popular startup IT hub", "", "", "",
-		5, recp.U.Index, contractor, originator, 4, 2, "blind")
-
-	if err != nil {
-		return err
-	}
-
-	_, err = testSolarProject(3, "260 1500 sq.ft homes set in a medieval cathedral style construction", 40000, "Trafalgar Square, London",
-		0, "Trafalgar Square is set in the heart of London's financial district, with big banks all over", "", "", "",
-		7, recp.U.Index, contractor, originator, 4, 2, "blind")
-
-	if err != nil {
-		return err
-	}
-
-	_, err = originator.Originate("100 16x24 panels on a solar rooftop", 14000, "Puerto Rico", 5, "ABC School in XYZ peninsula", 1, "blind") // 1 is the index for martin
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
